Fixed Repository Bug.  Size now long.  Updated BuildMain

+++ AZDORestApiExplorer.Domain/Git/Repository.cs
@@ -67,7 +67,7 @@ namespace AZDORestApiExplorer.Domain.Git
         public Project project { get; set; }
         public string remoteUrl { get; set; }
 
-        public int size { get; set; }
+        public long size { get; set; }
         public string sshUrl { get; set; }
         public string url { get; set; }
         public string webUrl { get; set; }
